requirements.txt
Django
Pillow
Pygments
argparse
django-registration-redux
django-simple-captcha
six
wsgiref
